requirements.txt
Django==3.2
djangorestframework==3.12.2
drf-yasg==1.20.0
psycopg2==2.9.1
djangorestframework-simplejwt==5.3.0
